requirements.txt
argparse
click
configparser
dnspython
grpcio
grpcio-tools
pkg-config
libvirt-python
lxml
os-win
oslo.concurrency
oslo.config
oslo.context
oslo.i18n
oslo.log
oslo.serialization
oslo.utils
Pillow
prettytable
protobuf
psutil
pystray
PyYAML
six
urllib3
wget
